Created new Logout component and linked to Navigation bar + removed initial logout from pages
